Move function to graphql utils
